Fixed issues 543, 551 and 552
