+++ src/components/core/blocks/Checkout/OrderReview.vue
@@ -6,10 +6,54 @@
 
 <script>
 export default {
-  name: 'order-review'
+  name: 'order-review',
+  props: ['isActive'],
+  data () {
+    return {
+      isFilled: false,
+      orderReview: {
+        terms: ''
+      }
+    }
+  },
+  methods: {
+    placeOrder () {
+      if (this.$store.state.checkout.personalDetails.createAccount) {
+        this.register()
+      } else {
+        this.$bus.$emit('checkout.placeOrder')
+      }
+    },
+    register () {
+      this.$bus.$emit('notification-progress-start', 'Registering the account ... ')
+      this.$store.dispatch('user/register', {
+        email: this.$store.state.checkout.personalDetails.emailAddress,
+        password: this.$store.state.checkout.personalDetails.password,
+        firstname: this.$store.state.checkout.personalDetails.firstName,
+        lastname: this.$store.state.checkout.personalDetails.lastName
+      }).then((result) => {
+        console.log(result)
+        this.$bus.$emit('notification-progress-stop')
+        if (result.code !== 200) {
+          this.$bus.$emit('notification', {
+            type: 'error',
+            message: result.result,
+            action1: { label: 'OK', action: 'close' }
+          })
+        } else {
+          this.$bus.$emit('notification', {
+            type: 'success',
+            message: 'You are logged in!',
+            action1: { label: 'OK', action: 'close' }
+          })
+          this.$store.commit('ui/setSignUp', false)
+          this.$bus.$emit('checkout.placeOrder')
+        }
+      }).catch(err => {
+        this.$bus.$emit('notification-progress-stop')
+        console.error(err)
+      })
+    }
+  }
 }
 </script>
-
-<style scoped>
-
-</style>
