+++ src/themes/default/components/core/blocks/Checkout/Shipping.vue
@@ -111,11 +111,7 @@
 </template>
 
 <script>
-import { mapState } from 'vuex'
 import { coreComponent } from 'lib/themes'
-import ShippingMethods from 'src/resource/shipping_methods.json'
-import Countries from 'src/resource/countries.json'
-
 import ButtonFull from 'theme/components/theme/ButtonFull.vue'
 import Tooltip from 'theme/components/core/Tooltip.vue'
 import { required, minLength } from 'vuelidate/lib/validators'
@@ -123,15 +119,6 @@ import { required, minLength } from 'vuelidate/lib/validators'
 // https://monterail.github.io/vuelidate/#sub-contextified-validators
 
 export default {
-  props: ['isActive'],
-  created () {
-    this.$bus.$on('checkout.personalDetails', (receivedData) => {
-      if (!this.isFilled) {
-        this.$store.dispatch('checkout/updatePropValue', ['firstName', receivedData.firstName])
-        this.$store.dispatch('checkout/updatePropValue', ['lastName', receivedData.lastName])
-      }
-    })
-  },
   validations: {
     shipping: {
       firstName: {
@@ -160,98 +147,6 @@ export default {
       city: {
         required
       }
-    }
-  },
-  data () {
-    return {
-      isFilled: false,
-      shippingMethods: ShippingMethods,
-      countries: Countries,
-      shipping: this.$store.state.checkout.shippingDetails,
-      shipToMyAddress: false,
-      myAddressDetails: {
-        firstname: '',
-        lastname: '',
-        country: '',
-        region: '',
-        city: '',
-        street: ['', ''],
-        postcode: '',
-        telephone: ''
-      }
-    }
-  },
-  computed: {
-    ...mapState({
-      currentUser: state => state.user.current
-    })
-  },
-  methods: {
-    sendDataToCheckout () {
-      this.$bus.$emit('checkout.shipping', this.shipping, this.$v)
-      this.isFilled = true
-    },
-    edit () {
-      if (this.isFilled) {
-        this.$bus.$emit('checkout.edit', 'shipping')
-        this.isFilled = false
-      }
-    },
-    hasShippingDetails () {
-      if (this.currentUser) {
-        if (this.currentUser.hasOwnProperty('default_shipping')) {
-          let id = this.currentUser.default_shipping
-          let addresses = this.currentUser.addresses
-          for (let i = 0; i < addresses.length; i++) {
-            if (addresses[i].id === Number(id)) {
-              this.myAddressDetails = addresses[i]
-              return true
-            }
-          }
-        }
-      }
-      return false
-    },
-    useMyAddress () {
-      this.shipToMyAddress = !this.shipToMyAddress
-      if (this.shipToMyAddress) {
-        this.shipping = {
-          firstName: this.myAddressDetails.firstname,
-          lastName: this.myAddressDetails.lastname,
-          country: this.myAddressDetails.country_id,
-          region: this.myAddressDetails.region.region ? this.myAddressDetails.region.region : '',
-          city: this.myAddressDetails.city,
-          streetAddress: this.myAddressDetails.street[0],
-          apartmentNumber: this.myAddressDetails.street[1],
-          zipCode: this.myAddressDetails.postcode,
-          phoneNumber: this.myAddressDetails.telephone,
-          shippingMethod: this.$store.state.checkout.shippingDetails.shippingMethod
-        }
-      } else {
-        this.shipping = this.$store.state.checkout.shippingDetails
-      }
-    },
-    getShippingMethod () {
-      for (let i = 0; i < ShippingMethods.length; i++) {
-        if (ShippingMethods[i].code === this.shipping.shippingMethod) {
-          return {
-            name: ShippingMethods[i].name,
-            cost: ShippingMethods[i].cost
-          }
-        }
-      }
-      return {
-        name: '',
-        cost: ''
-      }
-    },
-    getCountryName () {
-      for (let i = 0; i < this.countries.length; i++) {
-        if (this.countries[i].code === this.shipping.country) {
-          return this.countries[i].name
-        }
-      }
-      return ''
     }
   },
   components: {
