+++ src/themes/default/components/core/blocks/Header/AccountIcon.vue
@@ -1,6 +1,6 @@
 <template>
-  <div class="account-icon inline-flex">
-    <div class="dropdown" @mouseover="hoverIcon" @click="openMyAccount">
+  <div class="account-icon inline-flex" @click="openMyAccount">
+    <div class="dropdown" @mouseover="hoverIcon">
       <i class="material-icons md-18" @click="gotoAccount">account_circle</i>
       <div v-if="currentUser" :class="showDropdown ? 'dropdown-content show-dropdown' : 'dropdown-content'">
         <p>You're logged in as {{ currentUser.firstname }}</p>
